Feature #56 Add feedback for tree generation in Titan

+++ framework/commons/utils/src/main/java/com/puresoltechnologies/purifinity/framework/commons/utils/progress/AbstractProgressObservable.java
@@ -51,7 +51,7 @@ public class AbstractProgressObservable<Observable> implements
 		}
 	}
 
-	protected void fireStarted(String message, int totalAmount) {
+	protected void fireStarted(String message, long totalAmount) {
 		started = true;
 		this.totalAmount = totalAmount;
 		Iterator<WeakReference<ProgressObserver<Observable>>> iterator = observers
